rename /haka route to /saml
