+++ backend/src/routes.js
@@ -13,7 +13,7 @@ const skillLevels = require('./controllers/skill_level_controller.js')
 const assesmentResponse = require('./controllers/assesment_response_controller.js')
 const coursePersons = require('./controllers/course_person_controller.js')
 const grades = require('./controllers/grade_controller.js')
-const haka = require('./controllers/haka_controller.js')
+const saml = require('./controllers/saml_controller.js')
 
 const validateLang = require('./middleware/lang.js')
 const auth = require('./middleware/token_auth.js')
@@ -49,7 +49,7 @@ module.exports = (app) => {
   app.use(`${BASE_URL}/assesmentresponse`, assesmentResponse)
   app.use(`${BASE_URL}/selfassesment`, selfAssesment)
   app.use(`${BASE_URL}/login`, login)
-  app.use(`${BASE_URL}/haka`, haka)
+  app.use(`${BASE_URL}/saml`, saml)
   app.use(`${BASE_URL}/types`, types)
   app.use(`${BASE_URL}/skill-levels`, skillLevels)
   app.use(`${BASE_URL}/course-persons`, coursePersons)
